orangecrab_litedram/litedram_core.v: remove duplicated assignments
It seems that LiteX has problems with cases when a signal is assigned
to multiple wires like `assign {a, b} = c`, such line is repeated as
many times as there are signals on the left side. This causes warnings
(multiple conflicting drivers) if signal on the right side is inverted.

Signed-off-by: Krystian Hebel <krystian.hebel@3mdeb.com>