Makefile: make sure bootloader is rebuilt without marking it .PHONY
By marking a target .PHONY, it is always rebuilt, even when there
are no changes to the source code. As default target depends on
bootloader, it resulted in full rebuild on every invocation.

Now, thanks to empty prerequisite, 'make' is called recursively
for bootloader always, but the bootloader target won't be updated
if this isn't required. Top level 'make' invocation detects that
target was not changed and doesn't remake rest of the project in
such case, unless changes to other prerequisities were made.

Signed-off-by: Krystian Hebel <krystian.hebel@3mdeb.com>